Tweaking RStoXLSWriter
It's still horribly broken don't use it because it forces the close of
H2's connection to Y8 so as long as both of them are trying to use Y8
were going to be running into problems. I'm going to phase out Y8 in
this by using apache POI and I'm going to POI in the table engine so we
can prevent running a bunch of in memory databases.